requirements.txt
bcrypt==4.0.1
cffi==1.15.1
colour==0.1.5
cryptography==38.0.1
mccabe==0.7.0
paramiko==2.11.0
pycparser==2.21
pyflakes==2.5.0
PyNaCl==1.5.0
python-dotenv==0.21.0
PyYAML==6.0
six==1.16.0
toml==0.10.2
